pyproject.toml
[project]
name = "aiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "langchain-openrouter>=0.2.0",
    "langgraph>=1.1.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "transformers>=5.4.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]
